+++ pkg/clientset/versioned/typed/security/v1alpha1/generated_expansion.go
@@ -18,4 +18,6 @@ limitations under the License.
 
 package v1alpha1
 
+type AzureServicePrincipalExpansion interface{}
+
 type OrganizationExpansion interface{}